move: home 디렉토리->HomeSection으로 이름 변경
